Simplifying the Orchestrator bean by removing a useless layer
