--- alien4cloud-yorc-plugin/src/main/java/org/ystia/yorc/alien4cloud/plugin/YstiaOrchestrator.java
@@ -1,55 +1,0 @@
-/**
- * Copyright 2018 Bull S.A.S. Atos Technologies - Bull, Rue Jean Jaures, B.P.68, 78340, Les Clayes-sous-Bois, France.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.ystia.yorc.alien4cloud.plugin;
-
-import alien4cloud.orchestrators.plugin.ILocationConfiguratorPlugin;
-import alien4cloud.orchestrators.plugin.model.PluginArchive;
-import com.google.common.collect.Lists;
-import org.springframework.beans.factory.config.BeanDefinition;
-import org.springframework.context.annotation.Scope;
-import org.springframework.stereotype.Component;
-import org.ystia.yorc.alien4cloud.plugin.location.AbstractLocationConfigurerFactory;
-import org.ystia.yorc.alien4cloud.plugin.service.PluginArchiveService;
-
-import javax.inject.Inject;
-import java.util.List;
-
-/**
- * Yorc implementation for an orchestrator instance.
- */
-@Component
-@Scope(BeanDefinition.SCOPE_PROTOTYPE)
-public class YstiaOrchestrator extends YorcPaaSProvider {
-    @Inject
-    private AbstractLocationConfigurerFactory yorcLocationConfigurerFactory;
-
-    @Inject
-    private PluginArchiveService archiveService;
-
-    @Override
-    public ILocationConfiguratorPlugin getConfigurator(String locationType) {
-        return yorcLocationConfigurerFactory.newInstance(locationType);
-    }
-
-    @Override
-    public List<PluginArchive> pluginArchives() {
-        List<PluginArchive> archives = Lists.newArrayList();
-        archives.add(archiveService.parsePluginArchives("commons/resources"));
-        archives.add(archiveService.parsePluginArchives("docker/resources"));
-
-        return archives;
-    }
-}
